requirements.txt
Flask==3.1.0
opencv-python==4.9.0.80
numpy==1.26.4
keras==2.15.0
tensorflow-macos==2.15.0
google-generativeai==0.3.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1